scripts/training/requirements.txt
torch>=2.0.0
transformers>=4.30.0
peft>=0.4.0
datasets>=2.14.0
pillow>=9.5.0
accelerate>=0.21.0
scikit-learn>=1.3.0
numpy>=1.24.0
tqdm>=4.65.0
faiss-cpu>=1.7.4
